+++ packages/bugsnag_flutter/lib/src/client.dart
@@ -684,6 +684,7 @@ class Bugsnag extends BugsnagClient with DelegateClient {
     Map<String, Map<String, Object>>? metadata,
     List<BugsnagFeatureFlag>? featureFlags,
     List<BugsnagOnErrorCallback> onError = const [],
+    Set<BugsnagTelemetryType>? telemetry,
     Directory? persistenceDirectory,
     int? versionCode,
   }) async {
@@ -733,6 +734,9 @@ class Bugsnag extends BugsnagClient with DelegateClient {
       if (metadata != null) 'metadata': BugsnagMetadata(metadata),
       'featureFlags': featureFlags,
       'notifier': _notifier,
+      'telemetry': (telemetry ?? BugsnagTelemetryType.values)
+          .map((e) => e.toName())
+          .toList(),
       if (persistenceDirectory != null)
         'persistenceDirectory': persistenceDirectory.absolute.path,
       if (versionCode != null) 'versionCode': versionCode,