+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## TBD
 
+- Added `telemetry` option to `bugsnag.start` to allow sending of internal errors to be disabled.
 - Update bugsnag-android from v5.23.1 to [v5.25.0](https://github.com/bugsnag/bugsnag-android/blob/master/CHANGELOG.md#5250-2022-07-19)
 - Update bugsnag-cocoa from v6.18.1 to [v6.21.0](https://github.com/bugsnag/bugsnag-cocoa/blob/master/CHANGELOG.md#6210-2022-07-20)